test(fsio): make windows test at least regular file reads
